api/requirements.txt
aniso8601==9.0.1
click==7.1.2
docx2pdf==0.1.8
et-xmlfile==1.1.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.9.2
MarkupSafe==2.0.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
tqdm==4.64.1
Werkzeug==1.0.1
